chore: api: migrate route and controller
