+++ api/lib/application/sessions/session-controller.js
@@ -24,15 +24,6 @@ const getJuryCertificationSummaries = async function (
   return dependencies.juryCertificationSummarySerializer.serialize(juryCertificationSummaries, pagination);
 };
 
-const publish = async function (request, h, dependencies = { sessionManagementSerializer }) {
-  const sessionId = request.params.id;
-  const i18n = request.i18n;
-
-  const session = await usecases.publishSession({ sessionId, i18n });
-
-  return dependencies.sessionManagementSerializer.serialize({ session });
-};
-
 const publishInBatch = async function (request, h) {
   const sessionIds = request.payload.data.attributes.ids;
   const i18n = request.i18n;
@@ -65,7 +56,6 @@ const flagResultsAsSentToPrescriber = async function (request, h, dependencies =
 
 const sessionController = {
   getJuryCertificationSummaries,
-  publish,
   publishInBatch,
   unpublish,
   flagResultsAsSentToPrescriber,
