+++ api/scripts/certification/send-certification-result-emails.js
@@ -7,7 +7,7 @@ import i18n from 'i18n';
 
 import { disconnect } from '../../db/knex-database-connection.js';
 import * as mailService from '../../lib/domain/services/mail-service.js';
-import { manageEmails } from '../../lib/domain/services/session-publication-service.js';
+import { manageEmails } from '../../src/certification/session-management/domain/services/session-publication-service.js';
 import * as certificationCenterRepository from '../../src/certification/shared/infrastructure/repositories/certification-center-repository.js';
 import * as sharedSessionRepository from '../../src/certification/shared/infrastructure/repositories/session-repository.js';
 import { logger } from '../../src/shared/infrastructure/utils/logger.js';
