+++ api/src/certification/session-management/domain/services/mail-service.js
@@ -1,0 +1,95 @@
+import dayjs from 'dayjs';
+
+import * as translations from '../../../../../translations/index.js';
+import { config } from '../../../../shared/config.js';
+import { tokenService } from '../../../../shared/domain/services/token-service.js';
+import { mailer } from '../../../../shared/mail/infrastructure/services/mailer.js';
+
+const EMAIL_ADDRESS_NO_RESPONSE = 'ne-pas-repondre@pix.fr';
+
+// FRENCH_FRANCE
+const PIX_HOME_NAME_FRENCH_FRANCE = `pix${config.domain.tldFr}`;
+const PIX_HOME_URL_FRENCH_FRANCE = `${config.domain.pix + config.domain.tldFr}`;
+
+// INTERNATIONAL
+const PIX_HOME_NAME_INTERNATIONAL = `pix${config.domain.tldOrg}`;
+const PIX_HOME_URL_INTERNATIONAL = {
+  en: `${config.domain.pix + config.domain.tldOrg}/en/`,
+  fr: `${config.domain.pix + config.domain.tldOrg}/fr/`,
+  nl: `${config.domain.pix + config.domain.tldOrg}/nl-be/`,
+};
+
+function sendNotificationToCertificationCenterRefererForCleaResults({ email, sessionId, sessionDate }) {
+  const formattedSessionDate = dayjs(sessionDate).locale('fr').format('DD/MM/YYYY');
+
+  const options = {
+    from: EMAIL_ADDRESS_NO_RESPONSE,
+    fromName: translations.fr['email-sender-name']['pix-app'],
+    to: email,
+    template: mailer.acquiredCleaResultTemplateId,
+    variables: { sessionId, sessionDate: formattedSessionDate },
+  };
+
+  return mailer.sendEmail(options);
+}
+
+function sendCertificationResultEmail({
+  email,
+  sessionId,
+  sessionDate,
+  certificationCenterName,
+  resultRecipientEmail,
+  daysBeforeExpiration,
+  translate,
+}) {
+  const token = tokenService.createCertificationResultsByRecipientEmailLinkToken({
+    sessionId,
+    resultRecipientEmail,
+    daysBeforeExpiration,
+  });
+  const link = `${config.domain.pixApp + config.domain.tldOrg}/api/sessions/download-results/${token}`;
+
+  const formattedSessionDate = dayjs(sessionDate).locale('fr').format('DD/MM/YYYY');
+
+  const templateVariables = {
+    certificationCenterName,
+    sessionId,
+    sessionDate: formattedSessionDate,
+    fr: {
+      ...translations.fr['certification-result-email'].params,
+      homeName: PIX_HOME_NAME_FRENCH_FRANCE,
+      homeUrl: PIX_HOME_URL_FRENCH_FRANCE,
+      homeNameInternational: PIX_HOME_NAME_INTERNATIONAL,
+      homeUrlInternational: PIX_HOME_URL_INTERNATIONAL.fr,
+      title: translate({ phrase: 'certification-result-email.title', locale: 'fr' }, { sessionId }),
+      link: `${link}?lang=fr`,
+    },
+    en: {
+      ...translations.en['certification-result-email'].params,
+      homeName: PIX_HOME_NAME_INTERNATIONAL,
+      homeUrl: PIX_HOME_URL_INTERNATIONAL.en,
+      title: translate({ phrase: 'certification-result-email.title', locale: 'en' }, { sessionId }),
+      link: `${link}?lang=en`,
+    },
+  };
+
+  return mailer.sendEmail({
+    from: EMAIL_ADDRESS_NO_RESPONSE,
+    fromName: `${translations.fr['email-sender-name']['pix-app']} / ${translations.en['email-sender-name']['pix-app']}`,
+    to: email,
+    template: mailer.certificationResultTemplateId,
+    variables: templateVariables,
+  });
+}
+
+const mailService = {
+  sendNotificationToCertificationCenterRefererForCleaResults,
+  sendCertificationResultEmail,
+};
+
+/**
+ * @typedef {Object} MailService
+ * @property {function} sendCertificationResultEmail
+ * @property {function} sendNotificationToCertificationCenterRefererForCleaResults
+ */
+export { mailService, sendCertificationResultEmail, sendNotificationToCertificationCenterRefererForCleaResults };
