+++ api/src/certification/session-management/infrastructure/repositories/index.js
@@ -20,6 +20,7 @@ import * as certificationCandidateForSupervisingRepository from './certification
 import * as certificationCandidateRepository from './certification-candidate-repository.js';
 import * as certificationCompanionAlertRepository from './certification-companion-alert-repository.js';
 import * as certificationOfficerRepository from './certification-officer-repository.js';
+import * as certificationRepository from './certification-repository.js';
 import * as competenceMarkRepository from './competence-mark-repository.js';
 import * as courseAssessmentResultRepository from './course-assessment-result-repository.js';
 import * as cpfExportRepository from './cpf-export-repository.js';
@@ -55,6 +56,7 @@ import * as v3CertificationCourseDetailsForAdministrationRepository from './v3-c
  * @typedef {sessionRepository} SessionRepository
  * @typedef {supervisorAccessRepository} SupervisorAccessRepository
  * @typedef {certificationReportRepository} CertificationReportRepository
+ * @typedef {certificationRepository} CertificationRepository
  * @typedef {v3CertificationCourseDetailsForAdministrationRepository} V3CertificationCourseDetailsForAdministrationRepository
  * @typedef {competenceRepository} CompetenceRepository
  * @typedef {challengeRepository} ChallengeRepository
@@ -88,6 +90,7 @@ const repositoriesWithoutInjectedDependencies = {
   sessionRepository,
   sessionForSupervisingRepository,
   sessionJuryCommentRepository,
+  certificationRepository,
   certificationIssueReportRepository,
   supervisorAccessRepository,
   certificationReportRepository,
@@ -117,6 +120,7 @@ export {
   assessmentResultRepository,
   certificationChallengeRepository,
   certificationIssueReportRepository,
+  certificationRepository,
   challengeRepository,
   competenceMarkRepository,
   cpfExportRepository,
